refactor: add subtaask register
