+++ backend/helpers/pluginhelper/subtaskmeta_sorter/dependency.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package subTaskMetaSorter
+package subtaskmeta_sorter
 
 import (
 	"fmt"
