+++ backend/plugins/gitlab/impl/impl.go
@@ -19,7 +19,7 @@ package impl
 
 import (
 	"fmt"
-	"github.com/apache/incubator-devlake/helpers/pluginhelper/subTaskMetaSorter"
+	"github.com/apache/incubator-devlake/helpers/pluginhelper/subtaskmeta_sorter"
 	"time"
 
 	"github.com/apache/incubator-devlake/core/context"
@@ -49,7 +49,7 @@ type Gitlab string
 
 func init() {
 	// check subtask meta loop when init subtask meta
-	if _, err := subTaskMetaSorter.NewDependencySorter(tasks.SubTaskMetaList).Sort(); err != nil {
+	if _, err := subtaskmeta_sorter.NewDependencySorter(tasks.SubTaskMetaList).Sort(); err != nil {
 		panic(err)
 	}
 }
@@ -102,7 +102,7 @@ func (p Gitlab) Description() string {
 }
 
 func (p Gitlab) SubTaskMetas() []plugin.SubTaskMeta {
-	list, err := subTaskMetaSorter.NewDependencySorter(tasks.SubTaskMetaList).Sort()
+	list, err := subtaskmeta_sorter.NewDependencySorter(tasks.SubTaskMetaList).Sort()
 	if err != nil {
 		panic(err)
 	}
